LayoutTests/fast/text-extraction/debug-text-extraction-autofill-button-expected.txt

root
	input,'text',placeholder='Username'
	input,'email',placeholder='Email'
	input,'password',placeholder='Password',secure
